pyproject.toml
[tool.poetry]
name = "swiftlint-autodetect"
version = "0.1.0"
description = "TODO"
authors = ["Jonathan Wight <jwight@mac.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
typer = "^0.4.0"
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
swiftlint-autodetect = "swiftlint_autodetect.main:app"
